Type fix.
CVS patchset: 2352
CVS date: 2004/01/29 14:26:08

Mercurial revision: 0a4da57938557445d08d03d9b81b816d9115ca47

+++ crm/common/msgutils.c
@@ -392,7 +392,7 @@ decode_hash_value(gpointer value, char **node, char **subsys)
 		FNRET(FALSE);
 	    }
 	    cl_log(LOG_INFO, "Decoded value: (%s:%d)", *subsys, 
-		   strlen(*subsys));
+		   (int)strlen(*subsys));
 	    FNRET(TRUE);
     }
     else if(char_value != NULL)
